Fix edit wrong task bug
When there are at least two tasks on the current page and the user
tries to edit the task that wasn't the top task, the top task would
be the one being edited.

This was caused by having all the Task elements on the page subscribed
to the "Toggle Edit Style" topic which adds the css editing-task class
to the task the user wants to edit.

When the user try to edit a task, all the Task elements would add the
editing-task class which meant the App would only select the top
task to edit which in most cases was not the task the user
wanted to edit.

To fix the bug, the adding and removing of the css class editing-task
logic was moved from the Task component to the TaskModal component.
The Task elements are no longer subscribed to the "Toggle Edit Style"
topic.The TaskModal component is subscribed to a "Display Edit Modal"
topic where the TaskModal will add the editing-task to one Task element.